// src/app/models/pokemon.ts
export class Pokemon {
    name: string | undefined;
    number: number | undefined;
    types: string[] | undefined;
    imageUrl: string | undefined;
    height: number | undefined;
    weight: number | undefined;
    abilities: string[] | undefined;
    description?: string;
    gifUrl?: string;
}